Cast the frame header offset rather than letting it convert
A span iterator advanced by a size_t converts unsigned to signed, which
clang accepts silently and MSVC reports under some warning settings. The
cast says what is meant and costs nothing.

+++ NeuronCore/Protocol.cpp
@@ -214,7 +214,7 @@ bool DecodeFrame(std::span<const std::byte> _bytes, Frame& _outFrame)
 
   _outFrame.type = static_cast<MessageType>(type);
   _outFrame.version = version;
-  _outFrame.payload.assign(_bytes.begin() + FRAME_HEADER_BYTES, _bytes.end());
+  _outFrame.payload.assign(_bytes.begin() + static_cast<std::ptrdiff_t>(FRAME_HEADER_BYTES), _bytes.end());
   return true;
 }
